+++ pymemtrace/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Paul Ross"""
 __email__ = 'apaulross@gmail.com'
-__version__ = '0.1.7'
+__version__ = '0.2.0rc0'